requirements.txt
streamlit
streamlit-cropper
Pillow
numpy
pandas
scikit-learn
ultralytics
opencv-python-headless
requests
